environmentd: move statement logging tests to their own file
`server.rs` had grown to 7585 lines, and the statement logging tests were a
contiguous 994-line block of it with their own harness. Moving them to
`statement_logging.rs` leaves both files at a size a reader can hold, and it
keeps later additions to either group from churning the other's diff.

This is a pure move. Every moved item is byte-identical to its previous form,
and the only edits are the imports each file now needs.
